+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,3 +1,4 @@
 export { accountsController } from './accounts-controller';
 export { connectivityController } from './connectivity-controller';
 export { keyringController } from './keyring-controller';
+export { storageService } from './storage-service';