+++ packages/wallet/src/types.ts
@@ -1,5 +1,6 @@
 import type { ConnectivityAdapter } from '@metamask/connectivity-controller';
 import { KeyringControllerOptions } from '@metamask/keyring-controller';
+import { StorageAdapter } from '@metamask/storage-service';
 import type { Json } from '@metamask/utils';
 
 import type {
@@ -18,7 +19,7 @@ export type WalletOptions = {
     unknown,
     unknown
   >[];
-  instanceOptions?: InstanceSpecificOptions;
+  instanceOptions: InstanceSpecificOptions;
 };
 
 export type InstanceSpecificOptions = {
@@ -28,5 +29,9 @@ export type InstanceSpecificOptions = {
   keyringController?: {
     encryptor?: GenericEncryptor;
     keyringBuilders?: KeyringControllerOptions['keyringBuilders'];
+    keyringV2Builders?: KeyringControllerOptions['keyringV2Builders'];
+  };
+  storageService: {
+    storage: StorageAdapter;
   };
 };